feat: groundwork
