+++ src/pages/Markets/constants.ts
@@ -1,0 +1,9 @@
+export enum MARKETS_CATEGORIES {
+  TRADING_VOLUME = 'tradingVolume',
+  MARKET_CAP = 'marketCap',
+  TRENDING = 'trending',
+  TOP_MOVERS = 'topMovers',
+  RECENTLY_ADDED = 'recentlyAdded',
+  ONE_CLICK_DEFI = 'oneClickDefiAssets',
+  THORCHAIN_DEFI = 'thorchainDefi',
+}
